api/requirements.txt
quart
google-cloud-firestore
hypercorn
pydantic
pydantic-ai
termcolor
